feat: videoPlayer hooks
