+++ src/components/VideoPlayer/VideoPlayer.tsx
@@ -1,6 +1,7 @@
 import React, { useRef, useEffect } from "react";
 import Hls from "hls.js";
 import LockedVideoImg from "../../images/lockedVideo.png";
+import { attachMediaToHls, handleTimeUpdate } from "helpers/hooks/videoPlayer";
 
 interface VideoPlayerProps {
   videoLink: string;
@@ -28,45 +29,16 @@ const VideoPlayer: React.FC<VideoPlayerProps> = ({
 
     if (Hls.isSupported()) {
       const hls = new Hls();
-      attachMediaToHls(hls, videoElement);
+      attachMediaToHls(hls, videoElement, videoLink, videoId);
     }
   }, [videoLink, videoId]);
-
-  const handleSavedTime = (
-    videoElement: HTMLVideoElement,
-    videoId: string
-  ): void => {
-    const savedTime = localStorage.getItem(videoId);
-
-    if (savedTime !== null) {
-      videoElement.currentTime = parseInt(savedTime);
-    }
-  };
-
-  const attachMediaToHls = (hls: Hls, videoElement: HTMLVideoElement) => {
-    hls.attachMedia(videoElement);
-    hls.on(Hls.Events.MEDIA_ATTACHED, () => {
-      hls.loadSource(videoLink);
-      hls.on(Hls.Events.MANIFEST_PARSED, () => {
-        handleSavedTime(videoElement, videoId);
-      });
-    });
-  };
-
-  const handleTimeUpdate = (): void => {
-    const currentTime = videoRef.current?.currentTime || 0;
-
-    if (currentTime !== 0) {
-      localStorage.setItem(videoId, String(currentTime));
-    }
-  };
 
   return (
     <video
       width="100%"
       ref={videoRef}
       controls
-      onTimeUpdate={handleTimeUpdate}
+      onTimeUpdate={() => handleTimeUpdate(videoRef, videoId)}
       poster={isUnlocked ? poster : LockedVideoImg}
     />
   );
